More work in progress
Moved back to a single overall tick, but many more ticks per second. This
makes for smoother "scrolling" of the various numbers and also appears to be
better for performance. Also, we don't have to keep track of lots of
individual timers.

Games now have unique names.

Added tests for name generation.

There is a display a log of all hires and game releases at the bottom.